Extracted facilitator into seperate folder
