+++ index.js
@@ -27,7 +27,7 @@ const EconomySetup = require('./lib/EconomySetup');
 const Contracts = require('./lib/Contracts');
 const StakeHelper = require('./lib/helpers/stake/gateway_composer/StakeHelper');
 const Staker = require('./lib/helpers/stake/gateway_composer/Staker');
-const Facilitator = require('./lib/helpers/stake/gateway_composer/Facilitator');
+const Facilitator = require('./lib/Facilitator');
 
 module.exports = {
   AbiBinProvider: AbiBinProvider,
