+++ lib/Facilitator/index.js
@@ -20,14 +20,14 @@
 
 'use strict';
 
-const StakeHelper = require('./StakeHelper');
+const StakeHelper = require('../helpers/stake/gateway_composer/StakeHelper');
 
 /**
  * Facilitator performs below tasks:
  * - approves bounty amount to GatewayComposer
  * - calls GatewayComposer.acceptStakeRequest
  */
-class Facilitator {
+class Index {
   /**
    * Facilitator constructor object.
    *
@@ -106,4 +106,4 @@ class Facilitator {
   }
 }
 
-module.exports = Facilitator;
+module.exports = Index;
